Baseline commit
These run, but there's a lot of stuff I want to refactor (e.g. a lot of the math stuff is hacked together right now) and more features to add (e.g. replace nops so you can go in and out of fly mode, A and D key stuff so you can fly on the right vector)